fix(ux): improve CI stability and refine haptic feedback implementation
- Move \`triggerHapticFeedback\` to \`src/lib/utils.ts\` to resolve bundling issues.
- Use \`import type\` for database entities in hooks and components to prevent accidental module execution during build-time.
- Harden runtime guards for haptic utility to ensure absolute compatibility with Edge runtimes and Cloudflare Workers.
- Remove environmental artifacts from \`next-env.d.ts\` to maintain framework compliance.
- Verify that build and type-check pass successfully in a clean environment.
