+++ src/hooks/useTaskManagement.ts
@@ -3,8 +3,8 @@
 import { useState, useEffect, useCallback, useMemo, useRef } from 'react';
 import { createLogger } from '@/lib/logger';
 import { fetchWithTimeout } from '@/lib/api-client';
-import { Task, Deliverable } from '@/lib/db';
-import { triggerHapticFeedback } from '@/lib/haptics';
+import type { Task, Deliverable } from '@/lib/db';
+import { triggerHapticFeedback } from '@/lib/utils';
 
 export interface DeliverableWithTasks extends Deliverable {
   tasks: Task[];
